@@ -1,25 +1,53 @@
 {
     "private": true,
     "scripts": {
-        "dev": "npm run development",
+        "webpack": "webpack --hide-modules --progress -p",
+        "dev": "webpack --hide-modules --watch",
         "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
         "watch": "npm run development -- --watch",
         "watch-poll": "npm run watch -- --watch-poll",
         "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
         "prod": "npm run production",
-        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
+        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --hide-modules --progress --production"
     },
     "devDependencies": {
-        "axios": "^0.19",
-        "bootstrap": "^4.1.0",
-        "cross-env": "^5.1",
-        "jquery": "^3.2",
-        "laravel-mix": "^4.0.7",
-        "lodash": "^4.17.13",
+        "@babel/core": "^7.6.2",
+        "@babel/preset-env": "^7.6.2",
+        "@fortawesome/fontawesome-free": "^5.11.2",
+        "animate.css": "^3.7.0",
+        "axios": "^0.18.0",
+        "babel": "^6.23.0",
+        "babel-cli": "^6.26.0",
+        "babel-core": "^6.26.3",
+        "babel-loader": "^8.0.6",
+        "babel-preset-env": "^1.7.0",
+        "bulma": "^0.7.5",
+        "cross-env": "^5.2.1",
+        "css-loader": "^2.1.1",
+        "file-loader": "^4.2.0",
+        "font-awesome": "^4.7.0",
+        "font-awesome-webpack": "0.0.5-beta.2",
+        "mini-css-extract-plugin": "^0.6.0",
+        "moment": "^2.24.0",
+        "node-loader": "^0.6.0",
+        "node-sass": "^4.12.0",
+        "ol": "^5.3.3",
+        "plotly.js": "^1.49.5",
         "popper.js": "^1.12",
         "resolve-url-loader": "^2.3.1",
-        "sass": "^1.15.2",
-        "sass-loader": "^7.1.0",
-        "vue": "^2.5.17"
-    }
+        "sass": "^1.22.12",
+        "sass-loader": "^7.3.1",
+        "turf": "^3.0.14",
+        "url-loader": "^1.1.2",
+        "vue": "^2.6.10",
+        "vue-color": "^2.7.0",
+        "vue-loader": "^15.7.1",
+        "vue-style-loader": "^4.1.2",
+        "vue-template-compiler": "^2.6.10",
+        "vuejs-datepicker": "^1.6.2",
+        "vuex": "^3.1.1",
+        "webpack": "^4.41.0",
+        "webpack-cli": "^3.3.9"
+    },
+    "dependencies": {}
 }
