@@ -1,41 +1,33 @@
-window._ = require('lodash');
+import  Vue from 'vue';
+// import Vuex from 'vuex';
+import axios from 'axios';
+import moment from 'moment';
+import Plotly from 'plotly.js/dist/plotly';
 
-/**
- * We'll load jQuery and the Bootstrap jQuery plugin which provides support
- * for JavaScript based Bootstrap features such as modals and tabs. This
- * code may be modified to fit the specific needs of your application.
- */
 
-try {
-    window.Popper = require('popper.js').default;
-    window.$ = window.jQuery = require('jquery');
+import bulmaVue from './plugins/bulma-vue/bulma-vue.js';
 
-    require('bootstrap');
-} catch (e) {}
 
-/**
- * We'll load the axios HTTP library which allows us to easily issue requests
- * to our Laravel back-end. This library automatically handles sending the
- * CSRF token as a header based on the value of the "XSRF" token cookie.
- */
+window.Vue = Vue;
+window.axios = axios;
+window.moment = moment;
 
-window.axios = require('axios');
+moment.locale('ru');
 
-window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
 
-/**
- * Echo exposes an expressive API for subscribing to channels and listening
- * for events that are broadcast by Laravel. Echo and event broadcasting
- * allows your team to easily build robust real-time web applications.
- */
+window.Plotly = Plotly;
 
-// import Echo from 'laravel-echo';
 
-// window.Pusher = require('pusher-js');
 
-// window.Echo = new Echo({
-//     broadcaster: 'pusher',
-//     key: process.env.MIX_PUSHER_APP_KEY,
-//     cluster: process.env.MIX_PUSHER_APP_CLUSTER,
-//     encrypted: true
-// });
+window.axios.defaults.headers.common={
+  'X-CSRF-TOKEN': document.head.querySelector("[name=csrf-token]").getAttribute("content"),
+  'X-Requested-With': 'XMLHttpRequest',
+  //'Access-Control-Allow-Credentials': 'true'
+};
+
+
+
+Vue.use(bulmaVue);
+
+
+import '../sass/app.scss';
