@@ -12,7 +12,7 @@
 */
 
 Route::get('/', function () {
-    return view('welcome');
+    return view('ui.index');
 });
 
 Route::get('/data', 'StationDataController@index');
